修复群名片变更事件数据类型错误

+++ Protocol_OneBot/Notice/GroupMemberCardChanged.cs
@@ -8,9 +8,9 @@
 
         public string post_type { get; set; }
 
-        public int group_id { get; set; }
+        public long group_id { get; set; }
 
-        public int user_id { get; set; }
+        public long user_id { get; set; }
 
         public string notice_type { get; set; }
